refactor(style_code): refactor style coding and folder stucture
