+++ src/app/[...notfound]/layout.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from "react";
+import "@total-typescript/ts-reset";
 
 export default function Layout({ children }: { children: ReactNode }) {
   return <div className="max-w-5xl mx-auto container">{children}</div>;
