+++ src/app/middleware.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
+import "@total-typescript/ts-reset";
 
 export default function middleware(req: NextRequest) {
   const path = req.nextUrl.pathname;
