+++ src/components/atoms/Paragraph.tsx
@@ -3,6 +3,7 @@
 import { ParagraphProps } from "@/interfaces";
 import { cn } from "@/utils";
 import { memo } from "react";
+import "@total-typescript/ts-reset";
 
 function Paragraph({ children, className, ...props }: ParagraphProps) {
   return (
