+++ src/components/atoms/index.ts
@@ -1,5 +1,2 @@
 export { default as Paragraph } from "./Paragraph";
 export { default as Heading } from "./Heading";
-export { default as Card } from "./Card";
-
-export * from "./Dropdown";
