+++ src/components/molecules/index.ts
@@ -1,0 +1,3 @@
+export { default as Card } from "./Card";
+
+export * from "./Dropdown";
